task 4 and 5 impl

+++ workspace/HW3/src/hw3/store/TicketStore.java
@@ -6,7 +6,8 @@ import java.util.HashMap;
 import java.util.LinkedHashMap;
 
 public class TicketStore {
-	public static HashMap<Integer, Ticket> tickets = new LinkedHashMap<Integer, Ticket>();
+	
+	public static HashMap<Long, Ticket> tickets = new LinkedHashMap<Long, Ticket>();
 	
 //	public static void initStore() {
 //		TicketStore.tickets = new LinkedHashMap<Integer, Ticket>();
